// src/popup/enums/identifiers.ts
export const enum Identifiers {
    TextArea = "textArea",
    InputWords = "inputWords",
    InputCharacters = "inputCharacters",
    InputCharactersWithoutSpaces = "inputCharactersWithoutSpaces",
    DarkSwitch = "darkSwitch",
    Menu = "menu",
    MenuButton = "menuButton",
    Title = "title",
    CloseButton = "close",
}